pyproject.toml
[project]
name = "xray-access-log-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dnspython>=2.7.0",
    "geoip2>=5.1.0",
    "ipaddress>=1.0.23",
    "loguru>=0.7.3",
    "matplotlib>=3.10.3",
]
